backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
google-genai>=1.0.0
